+++ R/predict.fHMM_model.R
@@ -17,8 +17,9 @@
 #' confidence intervals.
 #'
 #' @examples
-#' data(dax_model_3t)
+#' data("dax_model_3t")
 #' predict(dax_model_3t)
+#' 
 #' @export
 #' @importFrom stats qt qgamma
 
@@ -26,13 +27,13 @@ predict.fHMM_model <- function(object, ahead = 5, alpha = 0.05, ...) {
 
   ### check input
   if (!inherits(object,"fHMM_model")) {
-    stop("'object' must be of class 'fHMM_model'.")
+    stop("'object' must be of class 'fHMM_model'.", call. = FALSE)
   }
   if (!(length(ahead) == 1 && is_number(ahead, int = TRUE, pos = TRUE))) {
-    stop("'ahead' must be a positive integer")
+    stop("'ahead' must be a positive integer.", call. = FALSE)
   }
   if (!(length(alpha) == 1 && is_number(alpha, pos = TRUE) && alpha <= 1)) {
-    stop("'alpha' must be a numeric between 0 and 1.")
+    stop("'alpha' must be a numeric between 0 and 1.", call. = FALSE)
   }
 
   ### extract parameters
@@ -42,7 +43,7 @@ predict.fHMM_model <- function(object, ahead = 5, alpha = 0.05, ...) {
   sdds <- object$data$controls$sdds
 
   ### predict states
-  state_prediction <- matrix(NA, nrow = ahead, ncol = M)
+  state_prediction <- matrix(NA_real_, nrow = ahead, ncol = M)
   last_state <- tail(if (object$data$controls$hierarchy) object$decoding[, 1] else object$decoding, n = 1)
   state_prob <- replace(numeric(M), last_state, 1)
   for (i in 1:ahead) {
@@ -54,7 +55,7 @@ predict.fHMM_model <- function(object, ahead = 5, alpha = 0.05, ...) {
   if (object$data$controls$hierarchy) {
     for (s in 1:M) {
       state_prob <- rep(1 / N, N)
-      fs_state_prediction <- matrix(NA, nrow = ahead, ncol = N)
+      fs_state_prediction <- matrix(NA_real_, nrow = ahead, ncol = N)
       for (i in 1:ahead) {
         state_prob <- state_prob %*% par$Gammas_star[[s]]
         fs_state_prediction[i, ] <- state_prediction[i, s] * state_prob
@@ -66,7 +67,7 @@ predict.fHMM_model <- function(object, ahead = 5, alpha = 0.05, ...) {
   }
 
   ### predict data
-  data_prediction <- matrix(NA, nrow = ahead, ncol = 3)
+  data_prediction <- matrix(NA_real_, nrow = ahead, ncol = 3)
   props <- sort(c(alpha, 0.5, 1 - alpha))
   if (!object$data$controls$hierarchy) {
     if (sdds[[1]]$name == "t") {
@@ -74,6 +75,14 @@ predict.fHMM_model <- function(object, ahead = 5, alpha = 0.05, ...) {
         data_prediction[i, ] <- sapply(props, function(x) {
           state_prediction[i, ] %*%
             (stats::qt(p = x, df = par$dfs) * par$sigmas + par$mus)
+        })
+      }
+    }
+    if (sdds[[1]]$name == "lnorm") {
+      for (i in 1:ahead) {
+        data_prediction[i, ] <- sapply(props, function(x) {
+          state_prediction[i, ] %*%
+            stats::qlnorm(p = x, meanlog = par$mus, sdlog = par$sigmas)
         })
       }
     }
@@ -93,7 +102,17 @@ predict.fHMM_model <- function(object, ahead = 5, alpha = 0.05, ...) {
       for (i in 1:ahead) {
         data_prediction[i, ] <- sapply(props, function(x) {
           state_prediction[i, -(1:M)] %*%
-            (stats::qt(p = x, df = unlist(par$dfs_star)) * unlist(par$sigmas_star) + unlist(par$mus_star))
+            (stats::qt(p = x, df = unlist(par$dfs_star)) * 
+               unlist(par$sigmas_star) + unlist(par$mus_star))
+        })
+      }
+    }
+    if (sdds[[2]]$name == "lnorm") {
+      for (i in 1:ahead) {
+        data_prediction[i, ] <- sapply(props, function(x) {
+          state_prediction[i, -(1:M)] %*%
+            stats::qlnorm(p = x, meanlog = par$mus_star, 
+                          sdlog = par$sigmas_star)
         })
       }
     }