+++ R/simulate_data.R
@@ -11,7 +11,8 @@
 #'  \item the matrix of \code{time_points},
 #'  \item the matrix of the simulated \code{markov_chain},
 #'  \item the matrix of the simulated \code{data},
-#'  \item the vector of fine-scale chunk sizes \code{T_star} if \code{controls$hierarchy = TRUE}.
+#'  \item the vector of fine-scale chunk sizes \code{T_star} if
+#'        \code{controls$hierarchy = TRUE}.
 #' }
 #'
 #' @keywords
@@ -23,13 +24,15 @@ simulate_data <- function(controls, true_parameters, seed = NULL) {
 
   ### check inputs
   if (!inherits(controls, "fHMM_controls")) {
-    stop("'controls' is not of class 'fHMM_controls'.")
+    stop("'controls' is not of class 'fHMM_controls'.", call. = FALSE)
   }
   if (!inherits(true_parameters, "fHMM_parameters")) {
-    stop("'true_parameters' is not of class 'fHMM_parameters'.")
+    stop("'true_parameters' is not of class 'fHMM_parameters'.",
+      call. = FALSE
+    )
   }
   if (!controls$simulated) {
-    stop("'controls$simulated' is not 'TRUE'.")
+    stop("'controls$simulated' is not 'TRUE'.", call. = FALSE)
   }
 
   ### simulate data
@@ -49,10 +52,22 @@ simulate_data <- function(controls, true_parameters, seed = NULL) {
     )
     time_points <- 1:controls[["horizon"]][1]
   } else {
-    T_star <- compute_T_star(horizon = controls[["horizon"]], period = controls[["period"]])
-    markov_chain <- matrix(NA, nrow = controls[["horizon"]][1], ncol = max(T_star) + 1)
-    data <- matrix(NA, nrow = controls[["horizon"]][1], ncol = max(T_star) + 1)
-    time_points <- matrix(NA, nrow = controls[["horizon"]][1], ncol = max(T_star) + 1)
+    T_star <- compute_T_star(
+      horizon = controls[["horizon"]],
+      period = controls[["period"]]
+    )
+    markov_chain <- matrix(NA_real_,
+      nrow = controls[["horizon"]][1],
+      ncol = max(T_star) + 1
+    )
+    data <- matrix(NA_real_,
+      nrow = controls[["horizon"]][1],
+      ncol = max(T_star) + 1
+    )
+    time_points <- matrix(NA_real_,
+      nrow = controls[["horizon"]][1],
+      ncol = max(T_star) + 1
+    )
     markov_chain[, 1] <- simulate_markov_chain(
       Gamma = true_parameters$Gamma,
       T = controls[["horizon"]][1],
@@ -84,7 +99,10 @@ simulate_data <- function(controls, true_parameters, seed = NULL) {
         seed = seed + t,
         total_length = max(T_star)
       )
-      time_points[t, -1] <- c(time_points[t, 1] - 1 + (1:T_star[t]), rep(NA, max(T_star) - T_star[t]))
+      time_points[t, -1] <- c(
+        time_points[t, 1] - 1 + (1:T_star[t]),
+        rep(NA_integer_, max(T_star) - T_star[t])
+      )
     }
   }
 
@@ -106,8 +124,8 @@ simulate_data <- function(controls, true_parameters, seed = NULL) {
 #' @param markov_chain
 #' A numeric vector of states of a Markov chain.
 #' @param sdd
-#' The name of the state-dependent distribution, one of \code{"t"} and
-#' \code{"gamma"}.
+#' The name of the state-dependent distribution, one of \code{"t"},
+#' \code{"gamma"}, and \code{"lnorm"}.
 #' @param mus
 #' A vector of expected values.
 #' @param sigmas
@@ -123,46 +141,55 @@ simulate_data <- function(controls, true_parameters, seed = NULL) {
 #' @return
 #' A numeric vector of length \code{total_length}, where the first
 #' \code{length(markov_chain)} elements are numeric values and the last
-#' \code{total_length - length(markov_chain)} elements are \code{NA}.
+#' \code{total_length - length(markov_chain)} elements are \code{NA_real_}.
 #'
 #' @keywords
 #' internal
 #'
 #' @importFrom stats rt rgamma
 
-simulate_observations <- function(markov_chain, sdd, mus, sigmas, dfs = NULL,
-                                  seed = NULL, total_length = length(markov_chain)) {
+simulate_observations <- function(markov_chain, sdd, mus, sigmas, dfs = NULL, seed = NULL,
+                                  total_length = length(markov_chain)) {
 
   ### check inputs
   if (!all(is_number(markov_chain, int = TRUE, pos = TRUE))) {
-    stop("'markov_chain' must be a numberic vector of states of a Markov chain.")
+    stop("'markov_chain' must be a numberic vector of Markov chain states.",
+      call. = FALSE
+    )
   }
-  if (!(length(sdd) == 1 && sdd %in% c("t", "gamma"))) {
-    stop("'sdd' must be one of 't' or 'gamma'.")
+  if (!(length(sdd) == 1 && sdd %in% c("t", "gamma", "lnorm"))) {
+    stop("'sdd' must be one of 't' or 'gamma'.", call. = FALSE)
   }
   if (!all(is_number(mus))) {
-    stop("'mus' must be a numberic vector.")
+    stop("'mus' must be a numberic vector.", call. = FALSE)
   }
   if (!all(is_number(sigmas, pos = TRUE))) {
-    stop("'sigmas' must be a positive numeric vector.")
+    stop("'sigmas' must be a positive numeric vector.", call. = FALSE)
   }
   if (!(length(mus) == length(sigmas))) {
-    stop("'mus' and 'sigmas' must be of the same length.")
+    stop("'mus' and 'sigmas' must be of the same length.", call. = FALSE)
   }
   if (sdd != "t") {
     if (!is.null(dfs)) {
-      stop("'dfs' must only be specified if 'sdd' = 't'.")
+      stop("'dfs' must only be specified if 'sdd' = 't'.", call. = FALSE)
     }
   } else {
     if (is.null(dfs)) {
-      stop("'dfs' must be specified if 'sdd' = 't'.")
+      stop("'dfs' must be specified if 'sdd' = 't'.", call. = FALSE)
     }
     if (!(all(is_number(dfs, pos = TRUE)) && length(dfs) == length(mus))) {
-      stop("'dfs' must be a positive number vector of length equal to 'mus' and 'sigmas'.")
+      stop(paste(
+        "'dfs' must be a positive number vector of length equal to",
+        "'mus' and 'sigmas'."
+      ), call. = FALSE)
     }
   }
-  if (!is_number(total_length) || length(total_length) != 1 || total_length < length(markov_chain)) {
-    stop("'total_length' must be an integer greater or equal than 'length(markov_chain)'.")
+  if (!is_number(total_length) || length(total_length) != 1 ||
+    total_length < length(markov_chain)) {
+    stop(paste(
+      "'total_length' must be an integer greater or equal than",
+      "'length(markov_chain)'."
+    ), call. = FALSE)
   }
 
   ### set seed
@@ -179,12 +206,18 @@ simulate_observations <- function(markov_chain, sdd, mus, sigmas, dfs = NULL,
       observations[t] <- stats::rt(1, dfs[s]) * sigmas[s] + mus[s]
     }
     if (sdd == "gamma") {
-      observations[t] <- stats::rgamma(1, shape = mus[s]^2 / sigmas[s]^2, scale = sigmas[s]^2 / mus[s])
+      observations[t] <- stats::rgamma(1,
+        shape = mus[s]^2 / sigmas[s]^2,
+        scale = sigmas[s]^2 / mus[s]
+      )
+    }
+    if (sdd == "lnorm") {
+      observations[t] <- stats::rlnorm(1, meanlog = mus[s], sdlog = sigmas[s])
     }
   }
 
   ### append NA's
-  observations <- c(observations, rep(NA, total_length - T))
+  observations <- c(observations, rep(NA_real_, total_length - T))
 
   ### return observations
   return(observations)