+++ NEWS.md
@@ -1,3 +1,8 @@
+# fHMM 1.0.3
+
+* We added the following methods for an `fHMM_model` object: `AIC()`, `BIC()`, `logLik()`, `nobs()`, `npar()`, `residuals()`.
+* The log-normal distribution can now be estimated by setting `sdds = "lnorm"` in the `controls` object.
+
 # fHMM 1.0.2
 
 * Fixed bug in `reorder_states()` that did not order the fine-scale parameter sets when the coarse-scale order was changed.